requirements.txt
# Framework Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# API REST y Documentación
Flask-RESTX==1.3.0
flask-cors==4.0.0
flasgger==0.9.7.1

# Autenticación y Seguridad
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
werkzeug==3.0.1

# Panel de Administración
Flask-Admin==1.6.1

# Base de datos
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Celery y Redis
celery==5.3.4
redis==5.0.1

# Email
Flask-Mail==0.9.1

# Variables de entorno
python-dotenv==1.0.0

# Servidor WSGI
gunicorn==21.2.0

# Utilidades
python-dateutil==2.8.2
pytz==2023.3

# Validación
marshmallow==3.20.1
flask-marshmallow==0.15.0

# Testing (opcional)
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
